New Modify

+++ database/migrations/2024_06_06_154729_data_santri.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama');
             $table->string('alamat');
-            $table->integer('no_hp');
+            $table->bigInteger('no_hp');
             $table->unique('no_hp');
             $table->string('jenis_kelamin');
             $table->string('asal_ponpes');
